+++ Presentation.Web/Controllers/API/V2/External/ItSystemUsages/ItSystemUsageV2Controller.cs
@@ -5,19 +5,14 @@ using System.Linq;
 using System.Net;
 using System.Web.Http;
 using System.Web.Http.Results;
-using Core.ApplicationServices.Extensions;
-using Core.ApplicationServices.Model.Shared;
 using Core.ApplicationServices.Model.SystemUsage.Write;
 using Core.ApplicationServices.SystemUsage;
 using Core.ApplicationServices.SystemUsage.Write;
-using Core.DomainModel.ItSystem.DataTypes;
 using Core.DomainModel.ItSystemUsage;
-using Core.DomainModel.ItSystemUsage.GDPR;
 using Core.DomainServices.Queries;
 using Core.DomainServices.Queries.SystemUsage;
 using Infrastructure.Services.Types;
 using Presentation.Web.Controllers.API.V2.External.ItSystemUsages.Mapping;
-using Presentation.Web.Controllers.API.V2.Mapping;
 using Presentation.Web.Extensions;
 using Presentation.Web.Infrastructure.Attributes;
 using Presentation.Web.Models.API.V2.Request.SystemUsage;
@@ -25,7 +20,6 @@ using Presentation.Web.Models.API.V2.Request.Generic.Queries;
 using Presentation.Web.Models.API.V2.Request.Generic.Roles;
 using Presentation.Web.Models.API.V2.Response.SystemUsage;
 using Presentation.Web.Models.API.V2.Types.Shared;
-using Presentation.Web.Models.API.V2.Types.SystemUsage;
 using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
@@ -40,12 +34,18 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
         private readonly IItSystemUsageService _itSystemUsageService;
         private readonly IItSystemUsageResponseMapper _responseMapper;
         private readonly IItSystemUsageWriteService _writeService;
+        private readonly IItSystemUsageWriteModelMapper _writeModelMapper;
 
-        public ItSystemUsageV2Controller(IItSystemUsageService itSystemUsageService, IItSystemUsageResponseMapper responseMapper, IItSystemUsageWriteService writeService)
+        public ItSystemUsageV2Controller(
+            IItSystemUsageService itSystemUsageService, 
+            IItSystemUsageResponseMapper responseMapper, 
+            IItSystemUsageWriteService writeService,
+            IItSystemUsageWriteModelMapper writeModelMapper)
         {
             _itSystemUsageService = itSystemUsageService;
             _responseMapper = responseMapper;
             _writeService = writeService;
+            _writeModelMapper = writeModelMapper;
         }
 
         /// <summary>
@@ -211,7 +211,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    GeneralProperties = MapFullUpdateGeneralData(request)
+                    GeneralProperties = _writeModelMapper.MapGeneralDataUpdate(request)
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -237,7 +237,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    Roles = MapRoles(request)
+                    Roles = _writeModelMapper.MapRoles(request)
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -263,7 +263,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    KLE = MapKle(request)
+                    KLE = _writeModelMapper.MapKle(request)
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -289,7 +289,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    ExternalReferences = MapReferences(request).FromNullable()
+                    ExternalReferences = _writeModelMapper.MapReferences(request).FromNullable()
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -315,7 +315,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    Archiving = MapArchiving(request)
+                    Archiving = _writeModelMapper.MapArchiving(request)
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -341,7 +341,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    GDPR = MapGDPR(request)
+                    GDPR = _writeModelMapper.MapGDPR(request)
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -367,7 +367,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             return _writeService
                 .Update(systemUsageUuid, new SystemUsageUpdateParameters
                 {
-                    OrganizationalUsage = MapOrganizationalUsage(request)
+                    OrganizationalUsage = _writeModelMapper.MapOrganizationalUsage(request)
                 })
                 .Select(_responseMapper.MapSystemUsageDTO)
                 .Match(Ok, FromOperationError);
@@ -462,17 +462,17 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
         /// <param name="request"></param>
         /// <param name="b"></param>
         /// <returns></returns>
-        private static SystemUsageUpdateParameters CreateFullPOSTParameters(CreateItSystemUsageRequestDTO request)
+        private SystemUsageUpdateParameters CreateFullPOSTParameters(CreateItSystemUsageRequestDTO request)
         {
             return new SystemUsageUpdateParameters
             {
-                GeneralProperties = request.General.FromNullable().Select(MapFullCommonGeneralData),
-                OrganizationalUsage = request.OrganizationUsage.FromNullable().Select(MapOrganizationalUsage),
-                KLE = request.LocalKleDeviations.FromNullable().Select(MapKle),
-                ExternalReferences = request.ExternalReferences.FromNullable().Select(MapReferences),
-                Roles = request.Roles.FromNullable().Select(MapRoles),
-                GDPR = request.GDPR.FromNullable().Select(MapGDPR),
-                Archiving = request.Archiving.FromNullable().Select(MapArchiving)
+                GeneralProperties = request.General.FromNullable().Select(_writeModelMapper.MapGeneralData),
+                OrganizationalUsage = request.OrganizationUsage.FromNullable().Select(_writeModelMapper.MapOrganizationalUsage),
+                KLE = request.LocalKleDeviations.FromNullable().Select(_writeModelMapper.MapKle),
+                ExternalReferences = request.ExternalReferences.FromNullable().Select(_writeModelMapper.MapReferences),
+                Roles = request.Roles.FromNullable().Select(_writeModelMapper.MapRoles),
+                GDPR = request.GDPR.FromNullable().Select(_writeModelMapper.MapGDPR),
+                Archiving = request.Archiving.FromNullable().Select(_writeModelMapper.MapArchiving)
             };
         }
 
@@ -482,7 +482,7 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
         /// <param name="request"></param>
         /// <param name="b"></param>
         /// <returns></returns>
-        private static SystemUsageUpdateParameters CreateFullPUTParameters(UpdateItSystemUsageRequestDTO request)
+        private SystemUsageUpdateParameters CreateFullPUTParameters(UpdateItSystemUsageRequestDTO request)
         {
             var generalDataInput = request.General ?? new GeneralDataUpdateRequestDTO();
             var orgUsageInput = request.OrganizationUsage ?? new OrganizationUsageWriteRequestDTO();
@@ -493,173 +493,13 @@ namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages
             var archiving = request.Archiving ?? new ArchivingWriteRequestDTO();
             return new SystemUsageUpdateParameters
             {
-                GeneralProperties = generalDataInput.FromNullable().Select(MapFullUpdateGeneralData),
-                OrganizationalUsage = orgUsageInput.FromNullable().Select(MapOrganizationalUsage),
-                KLE = kleInput.FromNullable().Select(MapKle),
-                ExternalReferences = externalReferenceDataDtos.FromNullable().Select(MapReferences),
-                Roles = roles.FromNullable().Select(MapRoles),
-                GDPR = gdpr.FromNullable().Select(MapGDPR),
-                Archiving = archiving.FromNullable().Select(MapArchiving)
-            };
-        }
-        private static UpdatedSystemUsageGDPRProperties MapGDPR(GDPRWriteRequestDTO request)
-        {
-            return new UpdatedSystemUsageGDPRProperties
-            {
-                Purpose = request.Purpose.AsChangedValue(),
-                BusinessCritical = MapYesNoDontKnow(request.BusinessCritical),
-                HostedAt = MapEnumChoice(request.HostedAt, HostedAtMappingExtensions.ToHostedAt),
-                DirectoryDocumentation = MapLink(request.DirectoryDocumentation),
-                DataSensitivityLevels = MapEnumList(request.DataSensitivityLevels, SensitiveDataLevelMappingExtensions.ToSensitiveDataLevel),
-                SensitivePersonDataUuids = MapCrossReferences(request.SensitivePersonDataUuids),
-                RegisteredDataCategoryUuids = MapCrossReferences(request.RegisteredDataCategoryUuids),
-                TechnicalPrecautionsInPlace = MapYesNoDontKnow(request.TechnicalPrecautionsInPlace),
-                TechnicalPrecautionsApplied = MapEnumList(request.TechnicalPrecautionsApplied, TechnicalPrecautionMappingExtensions.ToTechnicalPrecaution),
-                TechnicalPrecautionsDocumentation = MapLink(request.TechnicalPrecautionsDocumentation),
-                UserSupervision = MapYesNoDontKnow(request.UserSupervision),
-                UserSupervisionDate = request.UserSupervisionDate.AsChangedValue(),
-                UserSupervisionDocumentation = MapLink(request.UserSupervisionDocumentation),
-                RiskAssessmentConducted = MapYesNoDontKnow(request.RiskAssessmentConducted),
-                RiskAssessmentConductedDate = request.RiskAssessmentConductedDate.AsChangedValue(),
-                RiskAssessmentResult = MapEnumChoice(request.RiskAssessmentResult, RiskLevelMappingExtensions.ToRiskLevel),
-                RiskAssessmentDocumentation = MapLink(request.RiskAssessmentDocumentation),
-                RiskAssessmentNotes = request.RiskAssessmentNotes.AsChangedValue(),
-                DPIAConducted = MapYesNoDontKnow(request.DPIAConducted),
-                DPIADate = request.DPIADate.AsChangedValue(),
-                DPIADocumentation = MapLink(request.DPIADocumentation),
-                RetentionPeriodDefined = MapYesNoDontKnow(request.RetentionPeriodDefined),
-                NextDataRetentionEvaluationDate = request.NextDataRetentionEvaluationDate.AsChangedValue(),
-                DataRetentionEvaluationFrequencyInMonths = request.DataRetentionEvaluationFrequencyInMonths.AsChangedValue()
-            };
-        }
-        private static UpdatedSystemUsageArchivingParameters MapArchiving(ArchivingWriteRequestDTO archiving)
-        {
-            return new UpdatedSystemUsageArchivingParameters()
-            {
-                ArchiveDuty = MapEnumChoice(archiving.ArchiveDuty, ArchiveDutyMappingExtensions.ToArchiveDutyTypes),
-                ArchiveTypeUuid = (archiving.ArchiveTypeUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue(),
-                ArchiveLocationUuid = (archiving.ArchiveLocationUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue(),
-                ArchiveTestLocationUuid = (archiving.ArchiveTestLocationUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue(),
-                ArchiveSupplierOrganizationUuid = (archiving.ArchiveSupplierOrganizationUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue(),
-                ArchiveActive = archiving.ArchiveActive.AsChangedValue(),
-                ArchiveNotes = archiving.ArchiveNotes.AsChangedValue(),
-                ArchiveFrequencyInMonths = archiving.ArchiveFrequencyInMonths.AsChangedValue(),
-                ArchiveDocumentBearing = archiving.ArchiveDocumentBearing.AsChangedValue(),
-                ArchiveJournalPeriods = archiving.ArchiveJournalPeriods.FromNullable().Select(periods => periods.Select(MapJournalPeriod)).AsChangedValue()
-            };
-        }
-
-        private static SystemUsageJournalPeriod MapJournalPeriod(JournalPeriodDTO journalPeriod)
-        {
-            return new SystemUsageJournalPeriod()
-            {
-                Approved = journalPeriod.Approved,
-                ArchiveId = journalPeriod.ArchiveId,
-                EndDate = journalPeriod.EndDate,
-                StartDate = journalPeriod.StartDate
-            };
-        }
-
-        private static ChangedValue<Maybe<IEnumerable<TOut>>> MapEnumList<TIn, TOut>(IEnumerable<TIn> list, Func<TIn, TOut> mapWith)
-        {
-            return (list?.FromNullable().Select(dataSensitivityLevelChoices => dataSensitivityLevelChoices.Select(mapWith)) ?? Maybe<IEnumerable<TOut>>.None).AsChangedValue();
-        }
-
-        private static ChangedValue<TOut?> MapEnumChoice<TIn, TOut>(TIn? choice, Func<TIn, TOut> mapWith) where TIn : struct where TOut : struct
-        {
-            return choice.HasValue ? mapWith(choice.Value) : new ChangedValue<TOut?>(null);
-        }
-
-        private static ChangedValue<Maybe<IEnumerable<Guid>>> MapCrossReferences(IEnumerable<Guid> crossReferences)
-        {
-            return (crossReferences?.FromNullable() ?? Maybe<IEnumerable<Guid>>.None).AsChangedValue();
-        }
-
-        private static ChangedValue<DataOptions?> MapYesNoDontKnow(YesNoDontKnowChoice? choice)
-        {
-            return MapEnumChoice(choice, DataOptionsMappingExtensions.ToDataOptions);
-        }
-
-        private static ChangedValue<Maybe<NamedLink>> MapLink(SimpleLinkDTO simpleLinkDto)
-        {
-            return (simpleLinkDto?.FromNullable().Select(linkDto => new NamedLink(linkDto.Name, linkDto.Url)) ?? Maybe<NamedLink>.None).AsChangedValue();
-        }
-
-        private static IEnumerable<UpdatedExternalReferenceProperties> MapReferences(IEnumerable<ExternalReferenceDataDTO> references)
-        {
-            return references.Select(x => new UpdatedExternalReferenceProperties
-            {
-                Title = x.Title,
-                DocumentId = x.DocumentId,
-                Url = x.Url,
-                MasterReference = x.MasterReference
-            });
-        }
-
-        private static UpdatedSystemUsageKLEDeviationParameters MapKle(LocalKLEDeviationsRequestDTO kle)
-        {
-            return new UpdatedSystemUsageKLEDeviationParameters
-            {
-                AddedKLEUuids = kle.AddedKLEUuids.FromNullable().AsChangedValue(),
-                RemovedKLEUuids = kle.RemovedKLEUuids.FromNullable().AsChangedValue()
-            };
-        }
-
-        /// <summary>
-        /// Maps a complete update of the general data update request. Nulled sections are interpreted as intentionally reset
-        /// </summary>
-        /// <param name="generalData"></param>
-        /// <returns></returns>
-        private static UpdatedSystemUsageGeneralProperties MapFullUpdateGeneralData(GeneralDataUpdateRequestDTO generalData)
-        {
-            var generalProperties = MapFullCommonGeneralData(generalData);
-            generalProperties.MainContractUuid = (generalData.MainContractUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue();
-            return generalProperties;
-        }
-
-        private static UpdatedSystemUsageOrganizationalUseParameters MapOrganizationalUsage(OrganizationUsageWriteRequestDTO input)
-        {
-            return new UpdatedSystemUsageOrganizationalUseParameters
-            {
-                ResponsibleOrganizationUnitUuid = (input.ResponsibleOrganizationUnitUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue(),
-                UsingOrganizationUnitUuids = (input.UsingOrganizationUnitUuids?.FromNullable() ?? Maybe<IEnumerable<Guid>>.None).AsChangedValue()
-            };
-        }
-
-        private static UpdatedSystemUsageRoles MapRoles(IEnumerable<RoleAssignmentRequestDTO> roles)
-        {
-            var roleAssignmentResponseDtos = roles.ToList();
-
-            return new UpdatedSystemUsageRoles
-            {
-                UserRolePairs = (roleAssignmentResponseDtos.Any() ?
-                    Maybe<IEnumerable<UserRolePair>>.Some(roleAssignmentResponseDtos.Select(x => new UserRolePair
-                    {
-                        RoleUuid = x.RoleUuid,
-                        UserUuid = x.UserUuid
-                    })) :
-                    Maybe<IEnumerable<UserRolePair>>.None).AsChangedValue()
-            };
-        }
-
-        private static UpdatedSystemUsageGeneralProperties MapFullCommonGeneralData(GeneralDataWriteRequestDTO generalData)
-        {
-            return new UpdatedSystemUsageGeneralProperties
-            {
-                LocalCallName = generalData.LocalCallName.AsChangedValue(),
-                LocalSystemId = generalData.LocalSystemId.AsChangedValue(),
-                Notes = generalData.Notes.AsChangedValue(),
-                SystemVersion = generalData.SystemVersion.AsChangedValue(),
-                DataClassificationUuid = (generalData.DataClassificationUuid?.FromNullable() ?? Maybe<Guid>.None).AsChangedValue(),
-                NumberOfExpectedUsersInterval = generalData
-                    .NumberOfExpectedUsers?
-                    .FromNullable()
-                    .Select(interval => (interval.LowerBound.GetValueOrDefault(0), interval.UpperBound)).Match(interval => interval, () => Maybe<(int, int?)>.None)
-                    .AsChangedValue(),
-                EnforceActive = ((generalData.Validity?.EnforcedValid)?.FromNullable() ?? Maybe<bool>.None).AsChangedValue(),
-                ValidFrom = (generalData.Validity?.ValidFrom?.FromNullable() ?? Maybe<DateTime>.None).AsChangedValue(),
-                ValidTo = (generalData.Validity?.ValidTo?.FromNullable() ?? Maybe<DateTime>.None).AsChangedValue(),
-                AssociatedProjectUuids = (generalData.AssociatedProjectUuids?.FromNullable() ?? Maybe<IEnumerable<Guid>>.None).AsChangedValue()
+                GeneralProperties = generalDataInput.FromNullable().Select(_writeModelMapper.MapGeneralDataUpdate),
+                OrganizationalUsage = orgUsageInput.FromNullable().Select(_writeModelMapper.MapOrganizationalUsage),
+                KLE = kleInput.FromNullable().Select(_writeModelMapper.MapKle),
+                ExternalReferences = externalReferenceDataDtos.FromNullable().Select(_writeModelMapper.MapReferences),
+                Roles = roles.FromNullable().Select(_writeModelMapper.MapRoles),
+                GDPR = gdpr.FromNullable().Select(_writeModelMapper.MapGDPR),
+                Archiving = archiving.FromNullable().Select(_writeModelMapper.MapArchiving)
             };
         }
