+++ Presentation.Web/Controllers/API/V2/External/ItSystemUsages/Mapping/IItSystemUsageWriteModelMapper.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections.Generic;
+using Core.ApplicationServices.Model.SystemUsage.Write;
+using Presentation.Web.Models.API.V2.Request.Generic.Roles;
+using Presentation.Web.Models.API.V2.Request.SystemUsage;
+using Presentation.Web.Models.API.V2.Types.Shared;
+
+namespace Presentation.Web.Controllers.API.V2.External.ItSystemUsages.Mapping
+{
+    public interface IItSystemUsageWriteModelMapper
+    {
+        UpdatedSystemUsageGDPRProperties MapGDPR(GDPRWriteRequestDTO request);
+        UpdatedSystemUsageArchivingParameters MapArchiving(ArchivingWriteRequestDTO archiving);
+        IEnumerable<UpdatedExternalReferenceProperties> MapReferences(IEnumerable<ExternalReferenceDataDTO> references);
+        UpdatedSystemUsageKLEDeviationParameters MapKle(LocalKLEDeviationsRequestDTO kle);
+        UpdatedSystemUsageGeneralProperties MapGeneralDataUpdate(GeneralDataUpdateRequestDTO generalData);
+        UpdatedSystemUsageOrganizationalUseParameters MapOrganizationalUsage(OrganizationUsageWriteRequestDTO input);
+        UpdatedSystemUsageRoles MapRoles(IEnumerable<RoleAssignmentRequestDTO> roles);
+        UpdatedSystemUsageGeneralProperties MapGeneralData(GeneralDataWriteRequestDTO generalData);
+    }
+}
